+++ backend/src/main/java/br/com/gtp/shared/constant/Messages.java
@@ -1,5 +1,0 @@
-package br.com.gtp.shared.constant;
-
-public class Messages {
-
-}