+++ backend/src/main/java/br/com/gtp/shared/enums/Gender.java
@@ -1,5 +1,8 @@
 package br.com.gtp.shared.enums;
 
 public enum Gender {
+	
+	MALE,
+    FEMALE
 
 }