+++ backend/src/main/java/br/com/gtp/shared/security/config/SecurityConfig.java
@@ -1,29 +1,30 @@
 package br.com.gtp.shared.security.config;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.http.HttpMethod;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.Customizer;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import br.com.gtp.shared.security.filter.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
 @RequiredArgsConstructor
 public class SecurityConfig {
 
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
     private final DaoAuthenticationProvider authenticationProvider;
 
     @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
 
         http
-
             .csrf(csrf -> csrf.disable())
 
             .sessionManagement(session ->
@@ -32,29 +33,38 @@ public class SecurityConfig {
             .authenticationProvider(authenticationProvider)
 
             .authorizeHttpRequests(auth -> auth
-            	.requestMatchers(
-            		HttpMethod.POST, "/api/v1/users"
-            		    ).permitAll()	
+
+                .requestMatchers(HttpMethod.POST, "/api/v1/users")
+                .permitAll()
+
+                .requestMatchers("/api/v1/auth/**")
+                .permitAll()
 
                 .requestMatchers(
-                		"/api/v1/auth/**",
                         "/v3/api-docs/**",
                         "/swagger-ui/**",
-                        "/swagger-ui.html"
-                ).permitAll()
-                .anyRequest().authenticated())
+                        "/swagger-ui.html")
+                .permitAll()
 
-            .httpBasic(Customizer.withDefaults());
+                .anyRequest()
+                .authenticated())
+
+            .addFilterBefore(
+                    jwtAuthenticationFilter,
+                    UsernamePasswordAuthenticationFilter.class)
+
+            .httpBasic(httpBasic -> httpBasic.disable())
+
+            .formLogin(formLogin -> formLogin.disable());
 
         return http.build();
     }
 
     @Bean
-    AuthenticationManager authenticationManager(
+    public AuthenticationManager authenticationManager(
             AuthenticationConfiguration configuration)
             throws Exception {
 
         return configuration.getAuthenticationManager();
     }
-
 }